GH-54 Move snapshot_suites out of libraries/testing as it is not needed for libraries/testing
